refactor: Split `SortMergeJoin` into multiple modules
